adzmart_ecommerce_backend-develop/requirements/prod.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
pytz==2022.2.1
sqlparse==0.4.2
psycopg2==2.9.3
hypercorn
uvloop
django-anymail[sendgrid]==8.6
